Klageorkestrator versjon 2

+++ bidrag-beregn-barnebidrag/src/main/kotlin/no/nav/bidrag/beregn/barnebidrag/service/AldersjusteringOrchestrator.kt
@@ -431,7 +431,7 @@ class AldersjusteringOrchestrator(
             listOf(
                 BeregnGrunnlagVedtak(
                     gjelderBarnReferanse = søknadsbarn.referanse,
-                    vedtakId = vedtaksId.toLong(),
+                    vedtakId = vedtaksId,
                     vedtakInnhold = vedtak,
                 ),
             ),
